QOL improvements
The Plugin now looks for portals at the eye-level instead of the default feet-level. (Makes it so that if your head and by extension your screen goes through a portal, you actually go through the portal and teleport)

The players velocity gets set to 0 every time he tries to go through a portal, This is because if you only flew mildly fast, you would exit the portal before the game could manage to teleport you. (If you fly really fast, you will still fly through and not get teleported, but I see this as a feature)