added README and plugin icon
While the README should be good enough for a long time, the icon is dare I say dogshit. It's just a temporary icon for the upcoming spigot and modrinth pages and should be replaced if either the plugin gets a decent amount of downloads or if I get an idea on how to make a good icon.